Move tasks tools into same file as terminal
Fixes #259018
