+++ src/vs/workbench/contrib/terminal/terminal.all.ts
@@ -16,7 +16,6 @@ import './browser/terminalView.js';
 import '../terminalContrib/accessibility/browser/terminal.accessibility.contribution.js';
 import '../terminalContrib/autoReplies/browser/terminal.autoReplies.contribution.js';
 import '../terminalContrib/chatAgentTools/browser/terminal.chatAgentTools.contribution.js';
-import '../terminalContrib/chatAgentTools/browser/task.chatAgentTools.contribution.js';
 import '../terminalContrib/developer/browser/terminal.developer.contribution.js';
 import '../terminalContrib/environmentChanges/browser/terminal.environmentChanges.contribution.js';
 import '../terminalContrib/find/browser/terminal.find.contribution.js';
