+++ src/vs/workbench/contrib/terminalContrib/chatAgentTools/browser/terminal.chatAgentTools.contribution.ts
@@ -12,6 +12,9 @@ import { registerWorkbenchContribution2, WorkbenchPhase, type IWorkbenchContribu
 import { ILanguageModelToolsService, ToolDataSource } from '../../../chat/common/languageModelToolsService.js';
 import { GetTerminalOutputTool, GetTerminalOutputToolData } from './getTerminalOutputTool.js';
 import { RunInTerminalTool, RunInTerminalToolData } from './runInTerminalTool.js';
+import { CreateAndRunTaskTool, CreateAndRunTaskToolData } from './task/createAndRunTaskTool.js';
+import { GetTaskOutputTool, GetTaskOutputToolData } from './task/getTaskOutputTool.js';
+import { RunTaskTool, RunTaskToolData } from './task/runTaskTool.js';
 
 // #region Workbench contributions
 
@@ -24,6 +27,9 @@ class ChatAgentToolsContribution extends Disposable implements IWorkbenchContrib
 		@ILanguageModelToolsService toolsService: ILanguageModelToolsService,
 	) {
 		super();
+
+		// #region Terminal
+
 		const runInTerminalTool = instantiationService.createInstance(RunInTerminalTool);
 		this._register(toolsService.registerToolData(RunInTerminalToolData));
 		this._register(toolsService.registerToolImplementation(RunInTerminalToolData.id, runInTerminalTool));
@@ -32,12 +38,37 @@ class ChatAgentToolsContribution extends Disposable implements IWorkbenchContrib
 		this._register(toolsService.registerToolData(GetTerminalOutputToolData));
 		this._register(toolsService.registerToolImplementation(GetTerminalOutputToolData.id, getTerminalOutputTool));
 
-		const toolSet = this._register(toolsService.createToolSet(ToolDataSource.Internal, 'runCommands', 'runCommands', {
+		const runCommandsToolSet = this._register(toolsService.createToolSet(ToolDataSource.Internal, 'runCommands', 'runCommands', {
 			icon: ThemeIcon.fromId(Codicon.terminal.id),
 			description: localize('toolset.runCommands', 'Runs commands in the terminal')
 		}));
-		toolSet.addTool(RunInTerminalToolData);
-		toolSet.addTool(GetTerminalOutputToolData);
+		runCommandsToolSet.addTool(RunInTerminalToolData);
+		runCommandsToolSet.addTool(GetTerminalOutputToolData);
+
+		// #endregion
+
+		// #region Tasks
+
+		const runTaskTool = instantiationService.createInstance(RunTaskTool);
+		this._register(toolsService.registerToolData(RunTaskToolData));
+		this._register(toolsService.registerToolImplementation(RunTaskToolData.id, runTaskTool));
+
+		const getTaskOutputTool = instantiationService.createInstance(GetTaskOutputTool);
+		this._register(toolsService.registerToolData(GetTaskOutputToolData));
+		this._register(toolsService.registerToolImplementation(GetTaskOutputToolData.id, getTaskOutputTool));
+
+		const createAndRunTaskTool = instantiationService.createInstance(CreateAndRunTaskTool);
+		this._register(toolsService.registerToolData(CreateAndRunTaskToolData));
+		this._register(toolsService.registerToolImplementation(CreateAndRunTaskToolData.id, createAndRunTaskTool));
+
+		const runTasksToolSet = this._register(toolsService.createToolSet(ToolDataSource.Internal, 'runTasks', 'runTasks', {
+			description: localize('toolset.runTasks', 'Runs tasks and gets their output for your workspace')
+		}));
+		runTasksToolSet.addTool(RunTaskToolData);
+		runTasksToolSet.addTool(GetTaskOutputToolData);
+		runTasksToolSet.addTool(CreateAndRunTaskToolData);
+
+		// #endregion
 	}
 }
 registerWorkbenchContribution2(ChatAgentToolsContribution.ID, ChatAgentToolsContribution, WorkbenchPhase.AfterRestored);
